Avoid total timeout for streaming HTTP clients
