+++ src/core/providers/databricks/provider.rs
@@ -437,7 +437,7 @@ impl LLMProvider for DatabricksProvider {
                 ProviderError::authentication("databricks", "API key is required")
             })?;
 
-        let client = crate::core::http::outbound::default_outbound_client().clone();
+        let client = crate::core::providers::base::streaming_client();
         let response = client
             .post(&url)
             .header("Authorization", format!("Bearer {}", api_key))
