+++ src/core/providers/empower/provider.rs
@@ -104,7 +104,7 @@ crate::define_pooled_http_provider_with_hooks!(
                 ProviderError::authentication(PROVIDER_NAME, "API key is required")
             })?;
 
-            let client = crate::core::http::outbound::default_outbound_client().clone();
+            let client = crate::core::providers::base::streaming_client();
             let response = client
                 .post(&url)
                 .header("Authorization", format!("Bearer {}", api_key))
