+++ src/core/providers/github/provider.rs
@@ -264,7 +264,7 @@ impl LLMProvider for GitHubProvider {
 
         // Execute streaming request using reqwest directly for SSE
         let url = format!("{}/chat/completions", self.config.get_api_base());
-        let client = crate::core::http::outbound::default_outbound_client().clone();
+        let client = crate::core::providers::base::streaming_client();
         let response = client
             .post(&url)
             .header("Authorization", format!("Bearer {}", api_key))
