+++ src/core/providers/ollama/provider.rs
@@ -561,9 +561,7 @@ impl LLMProvider for OllamaProvider {
 
         // Use reqwest directly for streaming
         let url = self.config.get_chat_endpoint();
-        let mut req = crate::core::http::outbound::default_outbound_client()
-            .clone()
-            .post(&url);
+        let mut req = crate::core::providers::base::streaming_client().post(&url);
 
         // Add auth header if API key is set
         if let Some(api_key) = self.config.get_api_key() {
