+++ src/core/providers/replicate/provider.rs
@@ -337,7 +337,7 @@ impl LLMProvider for ReplicateProvider {
                     ProviderError::authentication("replicate", "API token required")
                 })?;
 
-            let client = crate::core::http::outbound::default_outbound_client().clone();
+            let client = crate::core::providers::base::streaming_client();
             let response = client
                 .get(stream_url)
                 .header("Authorization", format!("Token {}", api_key))
